+++ testar/src/org/testar/reporting/HTMLreporter.java
@@ -1,4 +1,4 @@
-package org.testar.reporting_proofofconcept;
+package org.testar.reporting;
 
 import org.apache.commons.lang.StringEscapeUtils;
 import org.testar.OutputStructure;
@@ -12,13 +12,17 @@ import java.util.ArrayList;
 import java.util.Set;
 import java.util.StringJoiner;
 
-public class PlainTextReporter implements Reporting
+public class HTMLreporter implements Reporting
 {
-    private PlainTextReportUtil plainTextReportUtil;
+    private HTMLreportUtil htmlReportUtil;
     private int innerLoopCounter = 0;
-    public PlainTextReporter(String fileName, boolean replay) //replay or record mode
+    
+    private final String openBlockContainer = "<div class='block' style='display:flex;flex-direction:column'>";
+    private final String closeBlockContainer = "</div>";
+    
+    public HTMLreporter(String fileName, boolean replay) //replay or record mode
     {
-        plainTextReportUtil = new PlainTextReportUtil(fileName);
+        htmlReportUtil = new HTMLreportUtil(fileName);
         
         startReport();
         if(replay)  addReplayHeading();
@@ -27,17 +31,27 @@ public class PlainTextReporter implements Reporting
     
     private void startReport()
     {
-        plainTextReportUtil.addHeading(1, "TESTAR execution sequence report");
+        StringJoiner joiner = new StringJoiner("\n");
+        joiner.add("function reverse(){");
+        joiner.add("let direction = document.getElementById('main').style.flexDirection;");
+        joiner.add("if(direction === 'column') document.getElementById('main').style.flexDirection = " + "'column-reverse';");
+        joiner.add("else document.getElementById('main').style.flexDirection = 'column';}");
+        
+        htmlReportUtil.addHeader("TESTAR execution sequence report", joiner.toString());
     }
     
     private void addReplayHeading()
     {
-        plainTextReportUtil.addHeading(2, "TESTAR replay sequence report for file " + ConfigTags.PathToReplaySequence);
+        htmlReportUtil.addHeading(1, "TESTAR replay sequence report for file " + ConfigTags.PathToReplaySequence);
     }
     
     private void addRecordHeading()
     {
-        plainTextReportUtil.addHeading(2, "TESTAR execution sequence report for sequence " + OutputStructure.sequenceInnerLoopCount);
+            htmlReportUtil.addHeading(1, "TESTAR execution sequence report for sequence " + OutputStructure.sequenceInnerLoopCount);
+            // HTML button to invoke reverse function
+            htmlReportUtil.addContent("<button id='reverseButton' onclick='reverse()'>Reverse order</button>");
+            // Initialize the main div container to apply the reverse order
+            htmlReportUtil.addContent("<div id='main' style='display:flex;flex-direction:column'>");
     }
     
     @Override
@@ -46,19 +60,21 @@ public class PlainTextReporter implements Reporting
         String imagePath = prepareScreenshotImagePath(state.get(Tags.ScreenshotPath));
         String concreteIDCustom = state.get(Tags.ConcreteIDCustom, "NoConcreteIdCustomAvailable");
         String abstractIDCustom = state.get(Tags.AbstractIDCustom, "NoAbstractIdCustomAvailable");
-        
-        plainTextReportUtil.addHeading(3, "State " + innerLoopCounter);
-        plainTextReportUtil.addHeading(5, "ConcreteIDCustom="+concreteIDCustom);
-        plainTextReportUtil.addHeading(5, "AbstractIDCustom="+abstractIDCustom);
-        
+
+        htmlReportUtil.addContent(openBlockContainer); // Open state block container
+        htmlReportUtil.addHeading(2, "State " + innerLoopCounter);
+        htmlReportUtil.addHeading(4, "ConcreteIDCustom="+concreteIDCustom);
+        htmlReportUtil.addHeading(4, "AbstractIDCustom="+abstractIDCustom);
+
         String altText = "screenshot: state=" + innerLoopCounter + ", ConcreteIDCustom="+concreteIDCustom+", AbstractIDCustom="+abstractIDCustom;
-        plainTextReportUtil.addParagraph("Image: " + imagePath + " - " + altText);
+        htmlReportUtil.addParagraph("<img src=\"" + imagePath + "\" alt=\"" + altText + "\">");
+        htmlReportUtil.addContent(closeBlockContainer); // Close state block container
         
         innerLoopCounter++;
-        plainTextReportUtil.writeToFile();
+        htmlReportUtil.writeToFile();
     }
-    
-    
+
+
     private String prepareScreenshotImagePath(String path)
     {
         if(path.contains("./output"))
@@ -79,7 +95,7 @@ public class PlainTextReporter implements Reporting
         if(action.get(Tags.Desc)!=null)
         {
             String escaped = StringEscapeUtils.escapeHtml(action.get(Tags.Desc));
-            joiner.add(escaped);
+            joiner.add("<b>"+ escaped +"</b>");
         }
         joiner.add(StringEscapeUtils.escapeHtml(action.toString()));
         joiner.add("ConcreteIDCustom="+action.get(Tags.ConcreteIDCustom, "NoConcreteIdCustomAvailable"));
@@ -93,36 +109,40 @@ public class PlainTextReporter implements Reporting
     @Override
     public void addActions(Set<Action> actions)
     {
-        plainTextReportUtil.addHeading(4, "Set of actions:");
-        
+        htmlReportUtil.addContent(openBlockContainer); // Open derived actions block container
+        htmlReportUtil.addHeading(4, "Set of actions:");
+    
         ArrayList<String> actionStrings = new ArrayList<>();
         for(Action action:actions)
             actionStrings.add(getActionString(action));
+        
+        htmlReportUtil.addList(false, actionStrings);
+        htmlReportUtil.addContent(closeBlockContainer); // Close derived actions block container
     
-        plainTextReportUtil.addList(false, actionStrings);
-    
-        plainTextReportUtil.writeToFile();
+        htmlReportUtil.writeToFile();
     }
     
     @Override
     public void addActionsAndUnvisitedActions(Set<Action> actions, Set<String> concreteIdsOfUnvisitedActions)
     {
+        htmlReportUtil.addContent(openBlockContainer); // Open derived actions block container
+        
         ArrayList<String> actionStrings = new ArrayList<>();
         if(actions.size()==concreteIdsOfUnvisitedActions.size())
         {
-            plainTextReportUtil.addHeading(5, "Set of actions (all unvisited - a new state):");
+            htmlReportUtil.addHeading(4, "Set of actions (all unvisited - a new state):");
             for(Action action : actions)
                 actionStrings.add(getActionString(action));
         }
         else if(concreteIdsOfUnvisitedActions.isEmpty())
         {
-            plainTextReportUtil.addHeading(5, "All actions have been visited, set of available actions:");
+            htmlReportUtil.addHeading(4, "All actions have been visited, set of available actions:");
             for(Action action : actions)
                 actionStrings.add(getActionString(action));
         }
         else
         {
-            plainTextReportUtil.addHeading(5, concreteIdsOfUnvisitedActions.size()+" out of "+actions.size()+" actions have not been visited yet:");
+            htmlReportUtil.addHeading(4, concreteIdsOfUnvisitedActions.size()+" out of "+actions.size()+" actions have not been visited yet:");
             for(Action action : actions)
             {
                 if(concreteIdsOfUnvisitedActions.contains(action.get(Tags.ConcreteIDCustom, "NoConcreteIdCustomAvailable")))
@@ -132,9 +152,10 @@ public class PlainTextReporter implements Reporting
                 }
             }
         }
-        plainTextReportUtil.addList(false, actionStrings);
+        htmlReportUtil.addList(false, actionStrings);
+        htmlReportUtil.addContent(closeBlockContainer); // Close derived actions block container
     
-        plainTextReportUtil.writeToFile();
+        htmlReportUtil.writeToFile();
     }
     
     @Override
@@ -143,35 +164,36 @@ public class PlainTextReporter implements Reporting
         String screenshotDir = prepareScreenshotImagePath(OutputStructure.screenshotsOutputDir);
         String stateConcreteIDCustom = state.get(Tags.ConcreteIDCustom, "NoConcreteIdCustomAvailable");
         String actionConcreteIDCustom = action.get(Tags.ConcreteIDCustom, "NoConcreteIdCustomAvailable");
-        
+
         String actionPath = screenshotDir + "/"
                             + OutputStructure.startInnerLoopDateString + "_" + OutputStructure.executedSUTname
                             + "_sequence_" + OutputStructure.sequenceInnerLoopCount
                             + "/" + stateConcreteIDCustom
                             + "_" + actionConcreteIDCustom + ".png";
-        
-        plainTextReportUtil.addHeading(3, "Selected Action "+innerLoopCounter+" leading to State "+innerLoopCounter);
-        
+    
+        htmlReportUtil.addContent(openBlockContainer); // Open executed action block container
+        htmlReportUtil.addHeading(2, "Selected Action "+innerLoopCounter+" leading to State "+innerLoopCounter);
+    
         String stateString = "ConcreteIDCustom="+actionConcreteIDCustom;
         if(action.get(Tags.Desc) != null)
         {
             String escaped = StringEscapeUtils.escapeHtml(action.get(Tags.Desc));
             stateString += " || " + escaped;
         }
-        plainTextReportUtil.addHeading(5, stateString);
-        
-        
+        htmlReportUtil.addHeading(4, stateString);
+    
+    
         if(actionPath.contains("./output"))
             actionPath = actionPath.replace("./output","..");
-        
+
         actionPath = actionPath.replace("\\", "/");
-        
+
         String altText = "screenshot: action, ConcreteIDCustom="+actionConcreteIDCustom;
-        
-        plainTextReportUtil.addParagraph("Image: " + actionPath + " - " + altText);
     
+        htmlReportUtil.addParagraph("<img src=\""+actionPath+"\" alt=\"" + altText + "\">");
+        htmlReportUtil.addContent(closeBlockContainer); // Close executed action block container
     
-        plainTextReportUtil.writeToFile();
+        htmlReportUtil.writeToFile();
     }
     
     @Override
@@ -180,18 +202,22 @@ public class PlainTextReporter implements Reporting
         String verdictInfo = verdict.info();
         if(verdict.severity() > Verdict.OK.severity())
             verdictInfo = verdictInfo.replace(Verdict.OK.info(), "");
-        
-        plainTextReportUtil.addHorizontalLine();
-        plainTextReportUtil.addHeading(3, "Test verdict for this sequence: " + verdictInfo);
-        plainTextReportUtil.addHeading(5, "Severity: " + verdict.severity());
     
-        plainTextReportUtil.appendToFileName("_" + verdict.verdictSeverityTitle() + "_poc_");
-        plainTextReportUtil.writeToFile();
+        htmlReportUtil.addContent(openBlockContainer); // Open verdict block container
+        htmlReportUtil.addHeading(2, "Test verdict for this sequence: " + verdictInfo);
+        htmlReportUtil.addHeading(4, "Severity: " + verdict.severity());
+        htmlReportUtil.addContent(closeBlockContainer); // Close verdict block container
+        
+        htmlReportUtil.appendToFileName("_" + verdict.verdictSeverityTitle() + "_poc_");
+        htmlReportUtil.writeToFile();
     }
     
     @Override
     public void finishReport()
     {
-        plainTextReportUtil.writeToFile();
+        htmlReportUtil.addContent("</div>"); // Close the main div container
+        htmlReportUtil.addFooter();
+    
+        htmlReportUtil.writeToFile();
     }
 }
