+++ testar/src/org/testar/reporting/PlainTextReporter.java
@@ -1,4 +1,4 @@
-package org.testar.reporting_proofofconcept;
+package org.testar.reporting;
 
 import org.apache.commons.lang.StringEscapeUtils;
 import org.testar.OutputStructure;
@@ -12,17 +12,13 @@ import java.util.ArrayList;
 import java.util.Set;
 import java.util.StringJoiner;
 
-public class HTMLreporter implements Reporting
+public class PlainTextReporter implements Reporting
 {
-    private HTMLreportUtil htmlReportUtil;
+    private PlainTextReportUtil plainTextReportUtil;
     private int innerLoopCounter = 0;
-    
-    private final String openBlockContainer = "<div class='block' style='display:flex;flex-direction:column'>";
-    private final String closeBlockContainer = "</div>";
-    
-    public HTMLreporter(String fileName, boolean replay) //replay or record mode
+    public PlainTextReporter(String fileName, boolean replay) //replay or record mode
     {
-        htmlReportUtil = new HTMLreportUtil(fileName);
+        plainTextReportUtil = new PlainTextReportUtil(fileName);
         
         startReport();
         if(replay)  addReplayHeading();
@@ -31,27 +27,17 @@ public class HTMLreporter implements Reporting
     
     private void startReport()
     {
-        StringJoiner joiner = new StringJoiner("\n");
-        joiner.add("function reverse(){");
-        joiner.add("let direction = document.getElementById('main').style.flexDirection;");
-        joiner.add("if(direction === 'column') document.getElementById('main').style.flexDirection = " + "'column-reverse';");
-        joiner.add("else document.getElementById('main').style.flexDirection = 'column';}");
-        
-        htmlReportUtil.addHeader("TESTAR execution sequence report", joiner.toString());
+        plainTextReportUtil.addHeading(1, "TESTAR execution sequence report");
     }
     
     private void addReplayHeading()
     {
-        htmlReportUtil.addHeading(1, "TESTAR replay sequence report for file " + ConfigTags.PathToReplaySequence);
+        plainTextReportUtil.addHeading(2, "TESTAR replay sequence report for file " + ConfigTags.PathToReplaySequence);
     }
     
     private void addRecordHeading()
     {
-            htmlReportUtil.addHeading(1, "TESTAR execution sequence report for sequence " + OutputStructure.sequenceInnerLoopCount);
-            // HTML button to invoke reverse function
-            htmlReportUtil.addContent("<button id='reverseButton' onclick='reverse()'>Reverse order</button>");
-            // Initialize the main div container to apply the reverse order
-            htmlReportUtil.addContent("<div id='main' style='display:flex;flex-direction:column'>");
+        plainTextReportUtil.addHeading(2, "TESTAR execution sequence report for sequence " + OutputStructure.sequenceInnerLoopCount);
     }
     
     @Override
@@ -60,21 +46,19 @@ public class HTMLreporter implements Reporting
         String imagePath = prepareScreenshotImagePath(state.get(Tags.ScreenshotPath));
         String concreteIDCustom = state.get(Tags.ConcreteIDCustom, "NoConcreteIdCustomAvailable");
         String abstractIDCustom = state.get(Tags.AbstractIDCustom, "NoAbstractIdCustomAvailable");
-
-        htmlReportUtil.addContent(openBlockContainer); // Open state block container
-        htmlReportUtil.addHeading(2, "State " + innerLoopCounter);
-        htmlReportUtil.addHeading(4, "ConcreteIDCustom="+concreteIDCustom);
-        htmlReportUtil.addHeading(4, "AbstractIDCustom="+abstractIDCustom);
-
+        
+        plainTextReportUtil.addHeading(3, "State " + innerLoopCounter);
+        plainTextReportUtil.addHeading(5, "ConcreteIDCustom="+concreteIDCustom);
+        plainTextReportUtil.addHeading(5, "AbstractIDCustom="+abstractIDCustom);
+        
         String altText = "screenshot: state=" + innerLoopCounter + ", ConcreteIDCustom="+concreteIDCustom+", AbstractIDCustom="+abstractIDCustom;
-        htmlReportUtil.addParagraph("<img src=\"" + imagePath + "\" alt=\"" + altText + "\">");
-        htmlReportUtil.addContent(closeBlockContainer); // Close state block container
+        plainTextReportUtil.addParagraph("Image: " + imagePath + " - " + altText);
         
         innerLoopCounter++;
-        htmlReportUtil.writeToFile();
+        plainTextReportUtil.writeToFile();
     }
-
-
+    
+    
     private String prepareScreenshotImagePath(String path)
     {
         if(path.contains("./output"))
@@ -95,7 +79,7 @@ public class HTMLreporter implements Reporting
         if(action.get(Tags.Desc)!=null)
         {
             String escaped = StringEscapeUtils.escapeHtml(action.get(Tags.Desc));
-            joiner.add("<b>"+ escaped +"</b>");
+            joiner.add(escaped);
         }
         joiner.add(StringEscapeUtils.escapeHtml(action.toString()));
         joiner.add("ConcreteIDCustom="+action.get(Tags.ConcreteIDCustom, "NoConcreteIdCustomAvailable"));
@@ -109,40 +93,36 @@ public class HTMLreporter implements Reporting
     @Override
     public void addActions(Set<Action> actions)
     {
-        htmlReportUtil.addContent(openBlockContainer); // Open derived actions block container
-        htmlReportUtil.addHeading(4, "Set of actions:");
-    
+        plainTextReportUtil.addHeading(4, "Set of actions:");
+        
         ArrayList<String> actionStrings = new ArrayList<>();
         for(Action action:actions)
             actionStrings.add(getActionString(action));
-        
-        htmlReportUtil.addList(false, actionStrings);
-        htmlReportUtil.addContent(closeBlockContainer); // Close derived actions block container
     
-        htmlReportUtil.writeToFile();
+        plainTextReportUtil.addList(false, actionStrings);
+    
+        plainTextReportUtil.writeToFile();
     }
     
     @Override
     public void addActionsAndUnvisitedActions(Set<Action> actions, Set<String> concreteIdsOfUnvisitedActions)
     {
-        htmlReportUtil.addContent(openBlockContainer); // Open derived actions block container
-        
         ArrayList<String> actionStrings = new ArrayList<>();
         if(actions.size()==concreteIdsOfUnvisitedActions.size())
         {
-            htmlReportUtil.addHeading(4, "Set of actions (all unvisited - a new state):");
+            plainTextReportUtil.addHeading(5, "Set of actions (all unvisited - a new state):");
             for(Action action : actions)
                 actionStrings.add(getActionString(action));
         }
         else if(concreteIdsOfUnvisitedActions.isEmpty())
         {
-            htmlReportUtil.addHeading(4, "All actions have been visited, set of available actions:");
+            plainTextReportUtil.addHeading(5, "All actions have been visited, set of available actions:");
             for(Action action : actions)
                 actionStrings.add(getActionString(action));
         }
         else
         {
-            htmlReportUtil.addHeading(4, concreteIdsOfUnvisitedActions.size()+" out of "+actions.size()+" actions have not been visited yet:");
+            plainTextReportUtil.addHeading(5, concreteIdsOfUnvisitedActions.size()+" out of "+actions.size()+" actions have not been visited yet:");
             for(Action action : actions)
             {
                 if(concreteIdsOfUnvisitedActions.contains(action.get(Tags.ConcreteIDCustom, "NoConcreteIdCustomAvailable")))
@@ -152,10 +132,9 @@ public class HTMLreporter implements Reporting
                 }
             }
         }
-        htmlReportUtil.addList(false, actionStrings);
-        htmlReportUtil.addContent(closeBlockContainer); // Close derived actions block container
+        plainTextReportUtil.addList(false, actionStrings);
     
-        htmlReportUtil.writeToFile();
+        plainTextReportUtil.writeToFile();
     }
     
     @Override
@@ -164,36 +143,35 @@ public class HTMLreporter implements Reporting
         String screenshotDir = prepareScreenshotImagePath(OutputStructure.screenshotsOutputDir);
         String stateConcreteIDCustom = state.get(Tags.ConcreteIDCustom, "NoConcreteIdCustomAvailable");
         String actionConcreteIDCustom = action.get(Tags.ConcreteIDCustom, "NoConcreteIdCustomAvailable");
-
+        
         String actionPath = screenshotDir + "/"
                             + OutputStructure.startInnerLoopDateString + "_" + OutputStructure.executedSUTname
                             + "_sequence_" + OutputStructure.sequenceInnerLoopCount
                             + "/" + stateConcreteIDCustom
                             + "_" + actionConcreteIDCustom + ".png";
-    
-        htmlReportUtil.addContent(openBlockContainer); // Open executed action block container
-        htmlReportUtil.addHeading(2, "Selected Action "+innerLoopCounter+" leading to State "+innerLoopCounter);
-    
+        
+        plainTextReportUtil.addHeading(3, "Selected Action "+innerLoopCounter+" leading to State "+innerLoopCounter);
+        
         String stateString = "ConcreteIDCustom="+actionConcreteIDCustom;
         if(action.get(Tags.Desc) != null)
         {
             String escaped = StringEscapeUtils.escapeHtml(action.get(Tags.Desc));
             stateString += " || " + escaped;
         }
-        htmlReportUtil.addHeading(4, stateString);
-    
-    
+        plainTextReportUtil.addHeading(5, stateString);
+        
+        
         if(actionPath.contains("./output"))
             actionPath = actionPath.replace("./output","..");
-
+        
         actionPath = actionPath.replace("\\", "/");
-
+        
         String altText = "screenshot: action, ConcreteIDCustom="+actionConcreteIDCustom;
+        
+        plainTextReportUtil.addParagraph("Image: " + actionPath + " - " + altText);
     
-        htmlReportUtil.addParagraph("<img src=\""+actionPath+"\" alt=\"" + altText + "\">");
-        htmlReportUtil.addContent(closeBlockContainer); // Close executed action block container
     
-        htmlReportUtil.writeToFile();
+        plainTextReportUtil.writeToFile();
     }
     
     @Override
@@ -202,22 +180,18 @@ public class HTMLreporter implements Reporting
         String verdictInfo = verdict.info();
         if(verdict.severity() > Verdict.OK.severity())
             verdictInfo = verdictInfo.replace(Verdict.OK.info(), "");
-    
-        htmlReportUtil.addContent(openBlockContainer); // Open verdict block container
-        htmlReportUtil.addHeading(2, "Test verdict for this sequence: " + verdictInfo);
-        htmlReportUtil.addHeading(4, "Severity: " + verdict.severity());
-        htmlReportUtil.addContent(closeBlockContainer); // Close verdict block container
         
-        htmlReportUtil.appendToFileName("_" + verdict.verdictSeverityTitle() + "_poc_");
-        htmlReportUtil.writeToFile();
+        plainTextReportUtil.addHorizontalLine();
+        plainTextReportUtil.addHeading(3, "Test verdict for this sequence: " + verdictInfo);
+        plainTextReportUtil.addHeading(5, "Severity: " + verdict.severity());
+    
+        plainTextReportUtil.appendToFileName("_" + verdict.verdictSeverityTitle() + "_poc_");
+        plainTextReportUtil.writeToFile();
     }
     
     @Override
     public void finishReport()
     {
-        htmlReportUtil.addContent("</div>"); // Close the main div container
-        htmlReportUtil.addFooter();
-    
-        htmlReportUtil.writeToFile();
+        plainTextReportUtil.writeToFile();
     }
 }
