+++ testar/src/org/testar/reporting/ReportManager.java
@@ -1,4 +1,4 @@
-package org.testar.reporting_proofofconcept;
+package org.testar.reporting;
 
 import org.testar.OutputStructure;
 import org.testar.monkey.alayer.Action;
