+++ testar/src/org/testar/reporting_depricated/TestCase.java
@@ -28,7 +28,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *******************************************************************************************************/
 
-package org.testar.reporting;
+package org.testar.reporting_depricated;
 
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
